Add testcase for post index redirect to show page
